+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 * Handle address type validation issue after adding permission to user. Fixes UIU-912.
 * Add missing permissions. Fixes UIU-1088.
 * Accomodate changes in the shape of `/loan-storage/loan-history` data. Refs UIU-1092.
+* Fix expiration date for sponsor and proxy. Fixes UIU-924.
 
 ## [2.23.0](https://github.com/folio-org/ui-users/tree/v2.23.0) (2019-06-12)
 [Full Changelog](https://github.com/folio-org/ui-users/compare/v2.22.0...v2.23.0)